+++ src/interface/export/export_python.cc
@@ -223,25 +223,25 @@ void ExportPython::writeParam(const std::string &name, const Parameter &param, s
 	}
 }
 
-size_t ExportPython::createInstance() noexcept
+int ExportPython::createInstance() noexcept
 {
 	file_ << "yi.createInstance()\n";
 	return current_instance_id_++;
 }
 
-bool ExportPython::addInstanceObject(size_t instance_id, std::string &&base_object_name) noexcept
+bool ExportPython::addInstanceObject(int instance_id, std::string &&base_object_name) noexcept
 {
 	file_ << "yi.addInstanceObject(" << instance_id << ", \"" << base_object_name << "\")\n"; //FIXME Should I use the variable name "instance_id" for export instead?
 	return true;
 }
 
-bool ExportPython::addInstanceOfInstance(size_t instance_id, size_t base_instance_id) noexcept
+bool ExportPython::addInstanceOfInstance(int instance_id, size_t base_instance_id) noexcept
 {
 	file_ << "yi.addInstanceOfInstance(" << instance_id << ", " << base_instance_id << ")\n"; //FIXME Should I use the variable name "instance_id" for export instead?
 	return true;
 }
 
-bool ExportPython::addInstanceMatrix(size_t instance_id, Matrix4 &&obj_to_world, float time) noexcept
+bool ExportPython::addInstanceMatrix(int instance_id, Matrix4 &&obj_to_world, float time) noexcept
 {
 	file_ << "yi.addInstanceMatrix(" << instance_id << ", "; //FIXME Should I use the variable name "instance_id" for export instead?
 	writeMatrix(obj_to_world, file_);
